Consolidates trait and dynamically extend PHPUnit
